refactor : all videos code
