--- src/utilities/filters-utils.jsx
@@ -1,5 +1,0 @@
-const getFilteredVideos = (products, category) => {
-  if (category === "All") return products;
-  return products.filter((product) => product.category === category);
-};
-export { getFilteredVideos };
